make source fetch as a product
