+++ lib/ruby_wasm/build_system/product.rb
@@ -1,5 +1,6 @@
 require "rake"
 require_relative "product/product"
+require_relative "product/ruby_source"
 require_relative "product/baseruby"
 require_relative "product/zlib"
 require_relative "product/libyaml"
